requirements.txt
flask
pandas
qrcode
python-barcode
Pillow
openpyxl
python-docx
reportlab
